requirements.txt
pandas~=2.0.3
pg8000~=1.30.3
prefect~=2.14.3
python-dotenv~=0.21.0
selenium~=4.15.2
SQLAlchemy~=1.4.0
xgboost~=2.0.2
streamlit~=1.28.2
requests~=2.31.0
pillow~=10.1.0
python-dateutil~=2.8.2
python-dateutil~=2.8.2
numpy~=1.26.2
pytz~=2023.3.post1
scikit-learn~=1.3.2
matplotlib~=3.8.2
randomname